botCria/requirements.txt
numpy
matplotlib
requests
python-telegram-bot
numpy
praw